crypto: move decryption verification state tests to a new file
